+++ js/utils.js
@@ -1,0 +1,16 @@
+const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+const getRandomUniqueInt = (min, max) => {
+  const uinqueIntArr = [];
+  return function () {
+    while (uinqueIntArr.length < max - min + 1) {
+      const randomInt = getRandomInt(min, max);
+      if (!uinqueIntArr.includes(randomInt)) {
+        uinqueIntArr.push(randomInt);
+        return randomInt;
+      }
+    }
+  };
+};
+
+export {getRandomInt, getRandomUniqueInt};
